Break get_field tests into appropriate files
